Gracefully recover from multiple labeled disks
In Ubuntu, we attempt to guess the disk label by looking at
/dev/disk/by-label. It's possible that this contains multiple label
links. This ends up breaking the script in a weird way because $LABEL is
not quoted when passed to the virtfs command.

Solution is to bail out if we detect more than one labeled disk. It also
supports overriding LABEL if you know what you're doing, or at least to
provide an escape hatch.